Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.drawing.style.textRendering.settings.worker-expected.txt


FAIL Testing basic functionalities of textRendering in Canvas assert_equals: ctx.textRendering === "Auto" (got [undefined], expected Auto[string]) expected (string) "Auto" but got (undefined) undefined
